refactor(shell): split MobileLayoutsPanel into focused components
Decompose the 1010-LOC MobileLayoutsPanel.tsx into 4 sibling files. The
main component is now 406 LOC and contains only the panel-level state
management + layout switching/sharing handlers; sub-components live in
siblings.

Extracted files:

- MobileLayoutsPanelParts.tsx (229) — stateless presentational pieces:
  ICON_PATHS map + SvgIcon, ActionSheet (bottom-sheet portal),
  ShareOptionButton, SwipeActionButton, ActiveLayoutActions (3-button
  row shown for the active layout), LoadingSpinner, PermissionSelect.
  All used by both the list rows and the cloud-share panel.
- MobileLayoutsListItem.tsx (172) — LayoutListItem component (swipe-
  action overlay + thumbnail + name/preview/date + active-row action
  bar) and the small `findEntry` lookup helper.
- MobileCloudSharePanel.tsx (258) — full cloud-share bottom-sheet
  flow: idle/share/loading/success/error states, plus permission
  toggle, copy, and delete-share interactions.

Public surface unchanged: only `MobileLayoutsPanel` is exported from
this folder, no consumer files needed updates.

MobileLayoutsPanel.tsx: 1010 LOC -> 406 LOC. Lint clean. The existing
MobileLayoutsPanel.test.tsx suite (1 file / 32 tests) passes (2.78s
targeted run).
